该版本为有Redis版优惠升级后端第一次提交

+++ src/main/java/com/example/backend/Controller/ReviewController.java
@@ -41,7 +41,7 @@ public class ReviewController {
     }
 
     @GetMapping("/product/{productId}")
-    public ResponseEntity<?> getReviewsByProductId(@PathVariable("productId") Integer productId,
+    public ResponseEntity<?> getReviewsByProductId(@PathVariable("productId") Long productId,
                                                 @RequestParam(defaultValue = "1") int currentPage,
                                                 @RequestParam(defaultValue = "10") int pageSize) {
         return reviewService.getReviewsByProductId(productId,currentPage, pageSize);
